% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertPySpectrumToRSpectra}
\alias{convertPySpectrumToRSpectra}
\title{Prototpye for conversion of Python Spectrum}
\usage{
convertPySpectrumToRSpectra(x, reference = import("matchms"))
}
\description{
Prototpye for conversion of Python Spectrum
}
