% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertRSpectraToPySpectrum}
\alias{convertRSpectraToPySpectrum}
\title{Prototype for conversion of R Spectra}
\usage{
convertRSpectraToPySpectrum(x, reference = import("matchms"))
}
\description{
Prototype for conversion of R Spectra
}
