% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendPython.R
\name{MsBackendPython}
\alias{MsBackendPython}
\title{A MS data backend for data stored in Python}
\description{
The \code{MsBackendPython} relies on the MS data being stored/represented in a
Python object.
}
\details{
The \code{MsBackendPython} object references to the Python object with the MS
data purely through its variable name in the special \code{py} variable from
the \emph{reticulate} package. It needs to be evaluated whether this will work
in all situations, e.g. also in parallel processing setups.
}
\note{
\itemize{
\item we might need a \emph{modification counter} to ensure data subsettings (at
least those performed in R) are recorded. Or should we just compare the
lengths? length at initiation and length in Python?
}
}
\author{
Johannes Rainer and the EuBIC hackathon team
}
