% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendPython.R
\name{pyspec_copy_on_replace}
\alias{pyspec_copy_on_replace}
\title{Copy Python MS data structure on MS data replacement operations}
\usage{
pyspec_copy_on_replace(x = logical())
}
\arguments{
\item{x}{\code{logical()}}
}
\value{
If \code{pyspec_copy_on_replace()} is called without providing an input
parameter, a \code{logical(1)} is returned whether the \emph{copy-on-replace}
option is enabled or not.
}
\description{
Assigning a \code{MsBackendPy} to another variable name in R with e.g. \code{a <- b}
with \code{b} being a \code{MsBackendPy} object results in two different \code{MsBackendPy}
instances in R that point however to the \strong{same} MS data structure in
Python. Changing thus one of the two \code{MsBackendPy} either by subset or data
replacement operations changes the shared MS data in Python and thus
inadvertedly also the data from any other \code{MsBackendPy} instance in R
pointing to the same variable in Python.

Setting \code{pyspec_copy_on_replace(TRUE)} \strong{before} any data replacement
operation on a \code{MsBackendPy} instance causes the full MS data in Python to
be copied to a \strong{new} Phython variable before replacing any values. Thus,
in situations described above, \code{a} and \code{b} would then point to two different
Python variables. After the subset or replacement operation has been
performed, the \emph{copy-on-replace} setting can again be disabled with
\code{pyspec_copy_on_replace(FALSE)}.

In general, if there are multiple \code{MsBackendPy} instances (or \code{Spectra}
objects) pointing to the same Python variable, then
\code{pyspec_copy_on_replace(TRUE)} should be called before one of the replacement
methods \verb{$<-}, \verb{intensity<-}, \verb{mz<-}, \verb{peaksData<-}, \verb{spectraData<-} or
\code{applyProcessing()} is called on one of them. After the operation,
\code{pyspec_copy_on_replace(FALSE)} should be called.

See examples below for details.

Calling \code{pyspec_copy_on_replace()} without \code{TRUE} or \code{FALSE} returns a
\code{logical(1)} whether \emph{copy-on-replace} is enabled or not.

The \emph{copy-on-replace} approach avoids data corruptions by accidental
modification of common Python MS data shared by different \code{Spectra} objects/
\code{MsBackendPy} instances. On the downside, it comes with a slighlty lower
performance (as the copy has to be cloned) and can result in potentially
multiple copies of the (same) MS data in Python. \emph{Copy-on-replace} should
thus be used with care.
}
\details{
For the name of the Python MS data variable a number is appended
to the original variable name, ensuring that no other Python variable with
the same name exists (to avoid replacing any other variable in Python). For
example, if the original Python variable's name is \code{"my_data"}, \code{"my_data_1"}
is used if there is no other variable with that name. If there is already
another Python variable \code{"my_data_1"}, then \code{"my_data_2"} is used instead.

\emph{copy-on-replace} can also be enabled using an R option (e.g.
\code{options(pyspec_copy_on_replace = TRUE)}) or System environment variable
(e.g. \code{Sys.setenv(pyspec_copy_on_replace = TRUE)}).
}
\examples{

## Is copy-on-replace enabled?
pyspec_copy_on_replace()

## Setting *copy-on-replace* to FALSE (the default)
pyspec_copy_on_replace(FALSE)

## Load an example Spectra object and change to use a `MsBackendPy`
library(Spectra)
library(MsBackendMgf)
fl <- system.file("extdata", "mgf", "test.mgf", package = "SpectriPy")
s <- Spectra(fl, source = MsBackendMgf())
s <- setBackend(s, MsBackendPy(), pythonVariableName = "mgf_data")

## `s` is now a `Spectra` object with the data in Python, in a variable
## `"mgf_data"`.
s

## DATA CORRUPTION

## To show how data corruption can happen, we next create a subset of
## `s` and assign that to new variable `s_sub`. Both point to the same
## data in Python
s_sub <- s[1:10]
s_sub

## Without any data manipulation, both variables are valid
intensity(s)
intensity(s_sub)

## However, any data manipulation on any of the two variables will affect
## the shared MS data in Python. Below we assign retention times to the
## data subset `s_sub`. This will cause the associated Python data in
## `"mgf_data"` to be subset to the first 10 spectra and the retention times
## be added.
s_sub$rtime <- 1:10 + 0.1
rtime(s_sub)

## Calling `s` throws an error; the data in Python was restricted to the
## first 10 spectra, so, `s` is no longer *valid*.

## AVOID DATA COPRRUPTION WIHT COPY ON REPLACE

## If any MS data needs to be updated, replaced or added during an analysis,
## it is suggested to enable the *copy-on-replace* option to ensure that
## also the MS data in Python gets copied. We repeat the above analysis
## after enabling *copy-on-replace*:
s <- Spectra(fl, source = MsBackendMgf())
s <- setBackend(s, MsBackendPy(), pythonVariableName = "mgf_data")

pyspec_copy_on_replace(TRUE)

## Subset `s` and add/replace retention times in the subset
s_sub <- s[1:10]
s_sub

s_sub$rtime <- 1:10 + 0.5

## `s` and `s_sub` are now pointing to two **different** variables in Python
s
s_sub

## Thus, the data from `s` was not changed
rtime(s)

## While the one from `s_sub` was
rtime(s_sub)

## Disable *copy-on-replace* after the data replacement
pyspec_copy_on_replace(FALSE)
}
\author{
Johannes Rainer
}
