% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{spectraVariableMapping,missing-method}
\alias{spectraVariableMapping,missing-method}
\title{Convert between Spectra and matchms variable names}
\usage{
\S4method{spectraVariableMapping}{missing}(object, ...)
}
\arguments{
\item{object}{ignored.}

\item{...}{ignored.}
}
\value{
named \code{character} vector with names being \code{Spectra} variable names
and values the corresponding names in \code{matchms}.
}
\description{
Convert \code{Spectra} spectra variable names to names for matchms spectrum
metadata. Conversion is based on this \href{https://github.com/matchms/matchms/blob/master/matchms/data/known_key_conversions.csv}{definition in matchms}
}
\examples{

spectraVariableMapping()
}
\author{
Johannes Rainer
}
