% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion}
\alias{conversion}
\title{Low-level functions to convert between Spectra and matchms Spectrum}
\arguments{
\item{x}{For \code{r_to_py.Spectra()}: \code{Spectra} object. For \code{pyspec_to_rspec()}:
a Python list of matchms Spectrum objects.}
}
\value{
For \code{r_to_py.Spectra()}: Python array of Spectrum objects, same
length as \code{x}. For \code{pyspec_to_rspec()}: \code{\link[Spectra:Spectra]{Spectra::Spectra()}} with the
converted spectra.
}
\description{
The \code{r_to_py.Spectra()} and \code{pyspec_to_rspec()} functions allow to convert
R \code{\link[Spectra:Spectra]{Spectra::Spectra()}} objects into
\href{https://github.com/matchms/matchms}{matchms} Python \code{matchms.Spectrum}
objects. These functions are designed for \strong{advanced users or developers}
who want/need to integrate Python/matchms functionality into R using
\emph{reticulate}. All other users should use the dedicated R functions within
this package that take care of running the Python code in the correct Python
environment.

\code{.SPECTRA_2_MATCHMS} provides a mapping of core \code{Spectra} variables, based on
this \href{https://github.com/matchms/matchms/blob/master/matchms/data/known_key_conversions.csv}{definition in matchms},
that is used in the \code{r_to_py.Spectra()} and \code{pyspec_to_rspec()} functions.
}
\author{
Michael Witting, Johannes Rainer, Wout Bittremieux
}
